apps/backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pydantic==2.9.2
python-dotenv==1.0.1

# AI/ML
openai==1.54.0
httpx==0.27.0

# Data Processing
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
aiofiles==24.1.0

# Document Generation
python-docx==1.1.2
reportlab==4.2.5
python-pptx==1.0.2
fpdf2
weasyprint==62.3
markdown==3.7

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
Pillow==11.0.0

# Database (Supabase)
supabase==2.9.0

# Payments
stripe==8.0.0

# PDF Text Extraction
PyMuPDF==1.24.0

# Web Content Parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# Code Execution Sandbox
e2b-code-interpreter==1.0.3

# Browser Automation (Playwright)
playwright==1.49.1

# WebSocket support (included in uvicorn[standard])
websockets==12.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# V4 Agent Orchestrator dependencies
uvloop>=0.19.0
tiktoken>=0.5.0
# V9 Enhancement dependencies
cryptography>=42.0.0
